# Breaking changes: increase performance