--- a/tests/test_json.rs
+++ b/tests/test_json.rs
@@ -1,5 +1,4 @@
-extern crate ndarray;
-extern crate rusticsom;
+#![cfg(feature = "serde-1")]
 
 use rusticsom::*;
 mod data;