@JayKickliter 2020-10-14
Note: please don't merge this until someone with domain expertise checks that `train_random` is returning sensible results. I was able to add a test with reference output for `train_batch` to make sure I didn't break anything, but I don't know how to test random training.